Moved filter and search logic to Core from Django app
