+++ graph_explorer/application/urls.py
@@ -7,5 +7,5 @@ urlpatterns = [
     path('', views.index),
     path('load_data_source', views.load_data_source, name='load_data_source'),
     path('search', views.search, name='search'),
-    path('filter', views.filter, name='filter'),
+    path('filter', views.filter_graph, name='filter'),
 ]
